requirements.txt
streamlit>=1.28.0
Pillow
pandas
numpy
scikit-learn
tensorflow
matplotlib
seaborn
